fix: phone placement fix v2
After further investigation, if an entity does not have a height or width value, it defaults to 0. This should be generic enough to work in most instances. The only case I can find where it may not work is when placed under key doors (which is fine since they'll be unlocked in the run).